requirements.txt
# Core dependencies
pika==1.3.2
pyyaml==6.0.1
python-json-logger==2.0.7

# ML and data processing
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# Web framework
flask==3.0.0
flask-cors==4.0.0

# Monitoring
prometheus-client==0.19.0

# Security
cryptography==41.0.7
PyJWT==2.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
requests==2.31.0
